+++ packages/eslint-plugin/test/rules/one-level/external-dependency-nodes-v7.spec.ts
@@ -30,6 +30,7 @@ const dependencyNodesSettings = {
     {
       // mock('source')
       selector: "CallExpression[callee.name=mock] > Literal",
+      name: "mock",
     },
   ],
 };